requirements.txt
tensorflow==2.12.0
python-chess
joblib
streamlit
